backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4
google-genai==0.3.0
